JavaScript templates updated
