+++ golem-templates/templates/js/js-app-common/.gitignore
@@ -1,0 +1,3 @@
+/dist
+/golem-temp
+/node_modules
