+++ golem-templates/templates/js/js-app-common/common-js/golem.yaml
@@ -1,0 +1,39 @@
+# Schema for IDEA:
+# $schema: https://schema.golem.cloud/app/golem/1.1.1/golem.schema.json
+# Schema for vscode-yaml
+# yaml-language-server: $schema=https://schema.golem.cloud/app/golem/1.1.1/golem.schema.json
+
+templates:
+  js:
+    build:
+    - command: npx jco stubgen wit-generated -o src/generated
+      sources:
+      - wit-generated
+      targets:
+      - src/generated
+      rmdirs:
+      - src/generated
+      mkdirs:
+      - src/generated
+    - command: npx rollup -c
+      sources:
+      - src
+      - ../../common-ts
+      - rollup.config.mjs
+      - tsconfig.js
+      targets:
+      - dist/main.js
+    - command: npx jco componentize -w wit-generated -o dist/{{ component_name | to_snake_case }}.wasm dist/main.js
+      sources:
+      - dist/main.js
+      targets:
+      - dist/{{ component_name | to_snake_case }}.wasm
+    sourceWit: wit
+    generatedWit: wit-generated
+    componentWasm: dist/{{ component_name | to_snake_case }}.wasm
+    linkedWasm: ../../golem-temp/components/{{ component_name | to_snake_case }}.wasm
+    clean:
+    - dist
+customCommands:
+  npm-install:
+  - command: npm install
