+++ golem-templates/templates/js/js-app-common/common-js/src/lib.ts
@@ -1,0 +1,3 @@
+export function example_common_function() {
+    return "hello common";
+}
