+++ golem-templates/templates/js/js-app-common/golem.yaml
@@ -1,0 +1,10 @@
+# Schema for IDEA:
+# $schema: https://schema.golem.cloud/app/golem/1.1.1/golem.schema.json
+# Schema for vscode-yaml
+# yaml-language-server: $schema=https://schema.golem.cloud/app/golem/1.1.1/golem.schema.json
+
+includes:
+- common-*/golem.yaml
+- components-*/*/golem.yaml
+witDeps:
+- wit/deps
