+++ golem-templates/templates/js/js-app-common/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "app",
+  "workspaces": [
+    "common-js/*/*",
+    "components-js/*/*"
+  ],
+  "dependencies": {
+    "@golemcloud/golem-ts": "1.1.2"
+  },
+  "devDependencies": {
+    "@golemcloud/componentize-js": "0.10.5-golem.4",
+    "@golemcloud/jco": "1.4.4-golem.2",
+    "@rollup/plugin-alias": "^5.1.1",
+    "@rollup/plugin-node-resolve": "^16.0.0",
+    "rollup": "^4.29.1"
+  }
+}
