+++ golem-templates/templates/js/js-app-common/wit/common.wit
@@ -1,0 +1,1 @@
+package common:root;
