+++ golem-templates/templates/js/js-app-component-default/components-js/component-name/golem.yaml
@@ -1,0 +1,8 @@
+# Schema for IDEA:
+# $schema: https://schema.golem.cloud/app/golem/1.1.1/golem.schema.json
+# Schema for vscode-yaml
+# yaml-language-server: $schema=https://schema.golem.cloud/app/golem/1.1.1/golem.schema.json
+
+components:
+  componentname:
+    template: js
