+++ golem-templates/templates/js/js-app-component-default/components-js/component-name/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "component-name"
+}
