+++ golem-templates/templates/js/js-app-component-default/components-js/component-name/src/main.js
@@ -1,0 +1,11 @@
+let state = BigInt(0);
+
+export const componentNameApi = {
+    add(value) {
+        console.log(`Adding ${value} to the counter`);
+        state += value;
+    },
+    get() {
+        return state;
+    }
+}
