+++ golem-templates/templates/js/js-app-component-default/metadata.json
@@ -1,0 +1,4 @@
+{
+  "description": "The default component template for JavaScript",
+  "appComponentGroup": "default"
+}
