+++ golem-templates/templates/js/js-app-component-example-fetch/components-js/.gitignore
@@ -1,0 +1,3 @@
+/*/dist
+/*/src/generated
+/*/wit-generated
