+++ golem-templates/templates/js/js-app-component-example-fetch/components-js/component-name/src/main.js
@@ -1,0 +1,14 @@
+import {asyncToSyncAsResult} from "@golemcloud/golem-ts";
+
+let result;
+
+export const componentNameApi = {
+    getLastResult() {
+        return JSON.stringify(result);
+    },
+    fetchJson(url) {
+        result = asyncToSyncAsResult(fetch(url).then(response => response.json()));
+        console.log(result);
+        return JSON.stringify(result);
+    },
+}
