+++ golem-templates/templates/js/js-app-component-wasi-http/components-js/.gitignore
@@ -1,0 +1,3 @@
+/*/dist
+/*/src/generated
+/*/wit-generated
