+++ golem-templates/templates/js/js-app-component-wasi-http/components-js/component-name/rollup.config.mjs
@@ -1,0 +1,3 @@
+import componentRollupConfig from "../../common-js/rollup.config.component.mjs";
+
+export default componentRollupConfig();
