+++ golem-templates/templates/js/js-app-component-wasi-http/metadata.json
@@ -1,0 +1,4 @@
+{
+  "description": "A JavaScript worker implementing wasi:http/incoming-handler. For use with the http-handler api gateway binding type",
+  "appComponentGroup": "default"
+}
