--- golem-templates/templates/js/js-default-minimal/metadata.json
@@ -1,8 +1,0 @@
-{
-  "description": "A simple stateful Golem worker implemented in JavaScript with no dependencies on external services",
-  "requiresAdapter": false,
-  "exclude": [
-    "node_modules",
-    "out"
-  ]
-}
