--- golem-templates/templates/js/js-default-minimal/wit/main.wit
@@ -1,12 +1,0 @@
-package pack:name;
-
-// See https://component-model.bytecodealliance.org/design/wit.html for more details about the WIT syntax
-
-interface api {
-  add: func(value: u64);
-  get: func() -> u64;
-}
-
-world component-name  {
-  export api;
-}
