--- golem-templates/templates/js/js-default-wasi-http/.gitignore
@@ -1,3 +1,0 @@
-node_modules/
-out/
-src/generated/
