--- golem-templates/templates/js/js-default-wasi-http/metadata.json
@@ -1,9 +1,0 @@
-{
-  "description": "A simple stateful Golem worker implemented in JavaScript with no dependencies on external services",
-  "requiresAdapter": false,
-  "requiresWASI": true,
-  "exclude": [
-    "node_modules",
-    "out"
-  ]
-}
