--- golem-templates/templates/js/js-default/.gitignore
@@ -1,3 +1,0 @@
-node_modules/
-out/
-wit-generated/
