--- golem-templates/templates/js/js-default/golem.yaml
@@ -1,31 +1,0 @@
-# Schema for IDEA:
-# $schema: https://schema.golem.cloud/app/golem/1.1.1/golem.schema.json
-# Schema for vscode-yaml
-# yaml-language-server: $schema=https://schema.golem.cloud/app/golem/1.1.1/golem.schema.json
-
-tempDir: out/golem-temp
-components:
-  pack:name:
-    sourceWit: wit
-    generatedWit: wit-generated
-    componentWasm: out/components/component_name.wasm
-    linkedWasm: out/linked-components/component_name_linked.wasm
-    build:
-    - command: npx rollup --config
-      sources:
-      - src
-      - wit-generated
-      targets:
-      - out/main.js
-    - command: npx jco componentize -w wit-generated -o out/components/component_name.wasm out/main.js
-      mkdirs:
-      - out/components
-      sources:
-      - out/main.js
-      targets:
-      - out/components/component_name.wasm
-    clean:
-    - src/generated
-    customCommands:
-      install:
-      - command: npm install
