--- golem-templates/templates/js/js-default/metadata.json
@@ -1,12 +1,0 @@
-{
-  "description": "A stateful Golem worker implemented in JavaScript with full access to WASI and the Golem runtime APIs",
-  "requiresAdapter": false,
-  "requiresGolemHostWIT": true,
-  "requiresWASI": true,
-  "instructions": "INSTRUCTIONS-app",
-  "exclude": [
-    "node_modules",
-    "out",
-    "wit-generated"
-  ]
-}
