--- golem-templates/templates/js/js-default/package.json
@@ -1,9 +1,0 @@
-{
-  "devDependencies": {
-    "@golemcloud/componentize-js": "0.10.5-golem.4",
-    "@golemcloud/golem-ts": "1.1.2",
-    "@golemcloud/jco": "1.4.4-golem.2",
-    "@rollup/plugin-node-resolve": "^15.2.3",
-    "rollup": "^4.18.0"
-  }
-}
