--- golem-templates/templates/js/js-default/wit/main.wit
@@ -1,36 +1,0 @@
-package pack:name;
-
-// See https://component-model.bytecodealliance.org/design/wit.html for more details about the WIT syntax
-
-interface api {
-  add: func(value: u64);
-  get: func() -> u64;
-}
-
-world component-name  {
-  import golem:api/host@1.1.5;
-  import golem:rpc/types@0.1.3;
-
-  import wasi:blobstore/blobstore;
-  import wasi:blobstore/container;
-  import wasi:cli/environment@0.2.0;
-  import wasi:clocks/wall-clock@0.2.0;
-  import wasi:clocks/monotonic-clock@0.2.0;
-  import wasi:filesystem/preopens@0.2.0;
-  import wasi:filesystem/types@0.2.0;
-  import wasi:http/types@0.2.0;
-  import wasi:http/outgoing-handler@0.2.0;
-  import wasi:io/error@0.2.0;
-  import wasi:io/poll@0.2.0;
-  import wasi:io/streams@0.2.0;
-  import wasi:keyvalue/eventual-batch@0.1.0;
-  import wasi:keyvalue/eventual@0.1.0;
-  import wasi:logging/logging;
-  import wasi:random/random@0.2.0;
-  import wasi:random/insecure@0.2.0;
-  import wasi:random/insecure-seed@0.2.0;
-  import wasi:sockets/ip-name-lookup@0.2.0;
-  import wasi:sockets/instance-network@0.2.0;
-
-  export api;
-}
