--- golem-templates/templates/js/js-example-fetch/.gitignore
@@ -1,3 +1,0 @@
-node_modules/
-out/
-src/generated/
