--- golem-templates/templates/js/js-example-fetch/metadata.json
@@ -1,13 +1,0 @@
-{
-  "description": "JavaScript example of using fetch with the golem-ts library",
-  "requiresAdapter": false,
-  "requiresGolemHostWIT": true,
-  "requiresWASI": true,
-  "exclude": [
-    "node_modules",
-    "out",
-    "src/generated",
-    "src/main.d.ts",
-    "src/main.js"
-  ]
-}
