--- golem-templates/templates/js/js-example-fetch/package.json
@@ -1,14 +1,0 @@
-{
-  "scripts": {
-    "build": "rollup --config",
-    "componentize": "npm run build && jco componentize -w wit -o out/component_name.wasm out/main.js",
-    "clean": "rm -rf out src/generated"
-  },
-  "devDependencies": {
-    "@golemcloud/componentize-js": "0.10.5-golem.4",
-    "@golemcloud/golem-ts": "1.1.2",
-    "@golemcloud/jco": "1.4.4-golem.2",
-    "@rollup/plugin-node-resolve": "^15.2.3",
-    "rollup": "^4.18.0"
-  }
-}
