--- golem-templates/templates/js/js-example-fetch/wit/main.wit
@@ -1,13 +1,0 @@
-package pack:name;
-
-// See https://component-model.bytecodealliance.org/design/wit.html for more details about the WIT syntax
-
-interface api {
-  get-last-result: func() -> string;
-  fetch-json: func(url: string) -> string;
-}
-
-world component-name  {
-  import golem:api/host@1.1.5;
-  export api;
-}
